set EV_VERSION to the one used < 2.6.38
Most of the demo images use a crappy version of ts_lib that checks exactly if
EV_VERSION matches, so the update to 2.6.38 broke all these demos which work
fine if the change is backed out. If we ever plan to go mainline, this change
has to be reverted, of course.